+++ projects/packages/identity-crisis/src/class-identity-crisis.php
@@ -28,7 +28,7 @@ class Identity_Crisis {
 	/**
 	 * Package Version
 	 */
-	const PACKAGE_VERSION = '0.8.42';
+	const PACKAGE_VERSION = '0.8.43';
 
 	/**
 	 * Instance of the object.