+++ projects/plugins/jetpack/extensions/blocks/anchor-fm/editor.js
@@ -1,13 +1,6 @@
-import { JetpackLogo } from '@automattic/jetpack-components';
-import { useAnalytics } from '@automattic/jetpack-shared-extension-utils';
-import { Button, PanelRow } from '@wordpress/components';
 import { dispatch } from '@wordpress/data';
-import { PluginPostPublishPanel } from '@wordpress/edit-post';
 import { __ } from '@wordpress/i18n';
-import { external, Icon } from '@wordpress/icons';
-import { registerPlugin } from '@wordpress/plugins';
 import { castArray } from 'lodash';
-import { useEffect, useCallback } from 'react';
 import '@wordpress/notices';
 import { waitForEditor } from '../../shared/wait-for-editor';
 import { basicTemplate, spotifyBadgeTemplate } from './templates';
@@ -41,52 +34,6 @@ async function setEpisodeTitle( { title } ) {
 	}
 	await waitForEditor();
 	dispatch( 'core/editor' ).editPost( { title } );
-}
-
-const ConvertToAudio = () => {
-	const { tracks } = useAnalytics();
-
-	useEffect( () => {
-		tracks.recordEvent( 'jetpack_editor_block_anchor_fm_post_publish_impression' );
-	}, [ tracks ] );
-	const handleClick = useCallback(
-		() => tracks.recordEvent( 'jetpack_editor_block_anchor_fm_post_publish_click' ),
-		[ tracks ]
-	);
-	return (
-		<PluginPostPublishPanel
-			className="anchor-post-publish-outbound-link"
-			icon={ <JetpackLogo showText={ false } height={ 16 } logoColor="#1E1E1E" /> }
-			title={ __( 'Convert to audio', 'jetpack' ) }
-		>
-			<PanelRow>
-				<p>
-					{ __(
-						'Seamlessly turn this post into a podcast episode with Anchor - and let readers listen to your post.',
-						'jetpack'
-					) }
-				</p>
-			</PanelRow>
-			<div
-				role="link"
-				className="post-publish-panel__postpublish-buttons"
-				tabIndex={ 0 }
-				onClick={ handleClick }
-				onKeyDown={ handleClick }
-			>
-				<Button variant="secondary" href="https://anchor.fm/wordpressdotcom" target="_top">
-					{ __( 'Create a podcast episode', 'jetpack' ) }{ ' ' }
-					<Icon icon={ external } className="anchor-post-publish-outbound-link__external_icon" />
-				</Button>
-			</div>
-		</PluginPostPublishPanel>
-	);
-};
-
-function showPostPublishOutboundLink() {
-	registerPlugin( 'anchor-post-publish-outbound-link', {
-		render: ConvertToAudio,
-	} );
 }
 
 function createEpisodeErrorNotice( params ) {
@@ -127,9 +74,6 @@ function initAnchor() {
 				break;
 			case 'insert-episode-template':
 				insertTemplate( { ...actionParams, tpl: 'basicEpisode' } );
-				break;
-			case 'show-post-publish-outbound-link':
-				showPostPublishOutboundLink();
 				break;
 			case 'set-episode-title':
 				setEpisodeTitle( actionParams );