+++ projects/packages/post-list/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.5] - 2023-04-10
+### Added
+- Add Jetpack Autoloader package suggestion. [#29988]
+
 ## [0.4.4] - 2023-02-20
 ### Changed
 - Minor internal updates.
@@ -78,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated the default columns displayed on the post and page list screens
 - Refactored thumbnail preview to function server side. All javascript removed.
 
+[0.4.5]: https://github.com/automattic/jetpack-post-list/compare/v0.4.4...v0.4.5
 [0.4.4]: https://github.com/automattic/jetpack-post-list/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/automattic/jetpack-post-list/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/automattic/jetpack-post-list/compare/v0.4.1...v0.4.2